docs: rename mda_widget to mda_demo
